fix transaction issue

+++ src/main/java/com/blueharvest/bank/dto/AccountRequestDTO.java
@@ -3,10 +3,12 @@ package com.blueharvest.bank.dto;
 import lombok.Builder;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @Builder
 public class AccountRequestDTO {
     private String customerId;
-    private double initialCredit;
+    private BigDecimal initialCredit;
     private String accountType;
 }
