+++ src/main/java/com/blueharvest/bank/model/Transaction.java
@@ -1,10 +1,13 @@
 package com.blueharvest.bank.model;
 
+import lombok.Builder;
 import lombok.Data;
+
+import java.math.BigDecimal;
 
 @Data
 public class Transaction {
     private String transactionId;
     private String accountId;
-    private double amount;
+    private BigDecimal amount;
 }
